% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_evaporation.R
\name{calculate_potential_evaporation_stats}
\alias{calculate_potential_evaporation_stats}
\title{Calculate stats of potential evaporation for geographical subset}
\usage{
calculate_potential_evaporation_stats(matrices, geo_mask)
}
\arguments{
\item{matrices}{matrices}

\item{geo_mask}{geo_mask}
}
\value{
???
}
\description{
Calculate stats of potential evaporation for geographical subset
}
