% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_shapes_germany.R
\name{check_or_download_shapes_germany}
\alias{check_or_download_shapes_germany}
\title{Check Local Availability or Download Shape Files for Germany}
\usage{
check_or_download_shapes_germany()
}
\value{
path to folder containing shape files
}
\description{
This function checks if shape files for Germany are available in the folder
`\%TEMP\%/R_kwb.dwd/gadm40_DEU_shp`. If not, the shape files are downloaded
from <"https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_DEU_shp.zip">
and unpacked into that folder.
}
