% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin_to_raster_file.R
\name{convert_bin_to_raster_file}
\alias{convert_bin_to_raster_file}
\title{Convert Binary Radolan File to Raster File}
\usage{
convert_bin_to_raster_file(bin_file, target_dir = NULL, dbg = TRUE)
}
\arguments{
\item{bin_file}{path to binary radolan file or vector of such paths}

\item{target_dir}{directory in which to create the raster file(s). By default
the files are created in a subfolder "rst" parallel to the folder in which
the first element of \code{bin_file} resides.}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Convert Binary Radolan File to Raster File
}
