% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_radolan_radial.R
\name{coordinates_to_EPSG_4326}
\alias{coordinates_to_EPSG_4326}
\title{Coordinates to EPSG-4326}
\usage{
coordinates_to_EPSG_4326(latitude, longitude)
}
\arguments{
\item{latitude}{latitude}

\item{longitude}{longitude}
}
\value{
data.frame with columns lon and lat with CRS 4326
}
\description{
Coordinates to EPSG-4326
}
\examples{
coordinates_to_EPSG_4326(latitude = 10, longitude = 50)
}
