% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_radolan_radial.R
\name{crop_radolan_radial}
\alias{crop_radolan_radial}
\title{Crop Radolan Data}
\usage{
crop_radolan_radial(radolan, longitude, latitude, radius = 10)
}
\arguments{
\item{radolan}{radolan data as returned by
\code{\link{read_relevant_years_radolan}}}

\item{longitude}{longitude}

\item{latitude}{latitude}

\item{radius}{distance from point (latitude/longitude) in km}
}
\description{
Crop Radolan Data
}
