% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gz_file_and_unzip.R
\name{download_gz_file_and_unzip}
\alias{download_gz_file_and_unzip}
\title{Download a .gz File and Unzip it}
\usage{
download_gz_file_and_unzip(url, download_dir = tempdir())
}
\arguments{
\item{download_dir}{path to folder into which to download and unzip the file}

\item{URL}{URL to .gz file}
}
\value{
path to the unzipped file
}
\description{
Download a .gz File and Unzip it
}
