% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_radolan.R
\name{download_radolan}
\alias{download_radolan}
\title{Download Radolan Files on DWD Server}
\usage{
download_radolan(resolution = "daily", export_dir = "data", ...)
}
\arguments{
\item{resolution}{temporal resolution, one of "daily" or "hourly" (default:
"daily")}

\item{export_dir}{export directory (default: "data" in current working
directory)}

\item{\dots}{arguments passed to \link{get_radolan_urls}, such as \code{start_daily},
\code{start_hourly}, \code{end_daily}, \code{end_hourly}.}
}
\value{
list with "daily_historical_urls" and "hourly_historical_urls"
}
\description{
Download Radolan Files on DWD Server
}
\examples{
\dontrun{download_radolan(resolution = "daily")}
}
\seealso{
\itemize{
\item \link{get_radolan_urls},
\item \link{extract_radolan_zip_files},
\item \link{radolan_raw_to_raster},
\item \link{list_extracted_radolan_files}.
}
}
