% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_radolan_zip_files.R
\name{extract_radolan_zip_files}
\alias{extract_radolan_zip_files}
\title{Extract radolan zip files}
\usage{
extract_radolan_zip_files(radolan_dir)
}
\arguments{
\item{radolan_dir}{radolan directory}
}
\value{
unizipped in \code{radolan_dir} subfolder "bin"
}
\description{
Extract radolan zip files
}
