% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_description.R
\name{find_description_files}
\alias{find_description_files}
\title{For a Given URL, Try to Find Description Files}
\usage{
find_description_files(url, full_names = FALSE)
}
\arguments{
\item{url}{URL to a data file at ftp server by Deutscher Wetterdienst}

\item{full_names}{logical. If \code{TRUE} the full URLs are returned,
otherwise only the file names. The default is \code{FALSE}.}
}
\value{
vector of character with file names of or full URLs to all files with
  extensions ".txt" or ".pdf" that are in the same folder as the file that
  \code{url} points to
}
\description{
For a Given URL, Try to Find Description Files
}
