% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_monthly_variable_for_region.R
\name{get_berlin_dwd_mask}
\alias{get_berlin_dwd_mask}
\title{Get geographical "stamp" for Berlin area}
\usage{
get_berlin_dwd_mask()
}
\value{
???
}
\description{
Get geographical "stamp" for Berlin area
}
\examples{
get_berlin_dwd_mask()
}
