% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_urls_metadata.R
\name{get_dwd_urls_metadata}
\alias{get_dwd_urls_metadata}
\title{Get URLs to DWD Metadata}
\usage{
get_dwd_urls_metadata()
}
\value{
list of character with each element representing a URL to a metadata
file provided by Deutscher Wetterdienst (DWT), e.g. files
"_Beschreibung_Stationen.txt" describing measurement stations.
}
\description{
Get URLs to DWD Metadata
}
