% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_grid_germany.R
\name{get_example_grid_germany}
\alias{get_example_grid_germany}
\title{Read an Example Raster File for Germany}
\usage{
get_example_grid_germany()
}
\description{
Read monthly potential evaporation of January 2022, just as an example.
}
\seealso{
\itemize{
\item \link{get_shapes_of_germany}.
}
}
