% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radolan_projection_string.R
\name{get_radolan_projection_string}
\alias{get_radolan_projection_string}
\title{Compose radolan projection string}
\usage{
get_radolan_projection_string()
}
\value{
default radolan projection string
}
\description{
Compose radolan projection string
}
\examples{
get_radolan_projection_string()

}
