% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_radolan.R
\name{get_radolan_urls}
\alias{get_radolan_urls}
\title{Get URLs to Available Radolan Files on DWD Server}
\usage{
get_radolan_urls(start_daily = "2006-10", start_hourly = "2005-06",
  end_daily = format(format = "\%Y-\%m", lubridate::rollback(Sys.Date(),
  roll_to_first = TRUE)), end_hourly = end_daily)
}
\arguments{
\item{start_daily}{month string (yyyy-mm) of first daily record.
Default: "2006-10"}

\item{start_hourly}{month string (yyyy-mm) of first hourly record.
Default: "2005-06"}

\item{end_daily}{month string (yyyy-mm) of last daily record. Defaults to the
current month.}

\item{end_hourly}{month string (yyyy-mm) of last hourly records. Defaults to
\code{end_daily}.}
}
\value{
list with "daily_historical_urls" and "hourly_historical_urls"
}
\description{
Get URLs to Available Radolan Files on DWD Server
}
