% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_grids_germany.R
\name{list_grids_germany}
\alias{list_grids_germany}
\title{Get URLs to Zipped Files Containing Grid Data for Germany}
\usage{
list_grids_germany(
  resolution,
  extension,
  variable,
  from = NULL,
  to = NULL,
  recursive = TRUE
)
}
\arguments{
\item{resolution}{one of "monthly", "daily"}

\item{variable}{variable for which to look for URLs. Must be one of
\code{kwb.dwd::list_url(kwb.dwd:::ftp_path_grids_germany(resolution))}}

\item{from}{optional. First month to be considered, as "yyyymm" string}

\item{to}{optional. Last month to be considered, as "yyyymm" string}

\item{recursive}{whether to list files recursively. Default: \code{TRUE}}
}
\description{
Get URLs to Zipped Files Containing Grid Data for Germany
}
