% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_grids_germany.R
\name{list_monthly_grids_germany_asc_gz}
\alias{list_monthly_grids_germany_asc_gz}
\title{Get URLs to Monthly Grids in Zipped ESRI-ascii-grid Format}
\usage{
list_monthly_grids_germany_asc_gz(
  variable,
  from = NULL,
  to = NULL,
  recursive = TRUE
)
}
\arguments{
\item{variable}{variable for which to look for URLs. Must be one of
\code{kwb.dwd::list_url(kwb.dwd:::ftp_path_grids_germany("monthly"))}}

\item{from}{optional. First month to be considered, as "yyyymm" string}

\item{to}{optional. Last month to be considered, as "yyyymm" string}

\item{recursive}{whether to list files recursively. Default: \code{TRUE}}
}
\description{
Get URLs to Monthly Grids in Zipped ESRI-ascii-grid Format
}
