% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_url.R
\name{list_url}
\alias{list_url}
\title{List Files or Folders in a FTP Directory}
\usage{
list_url(url, recursive = FALSE, n_attempts = 3)
}
\arguments{
\item{url}{URL to FTP server, including "ftp://"}

\item{recursive}{logical indicating whether to list files in all
subdirectories (default: \code{FALSE})}

\item{n_attempts}{number of attempts when trying to access an URL.}
}
\description{
List Files or Folders in a FTP Directory
}
