% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_zipped_esri_ascii_grids}
\alias{list_zipped_esri_ascii_grids}
\title{Get URLs of Files in Zipped ESRI-ascii-grid Format}
\usage{
list_zipped_esri_ascii_grids(base_url, recursive = TRUE)
}
\arguments{
\item{base_url}{URL from which to start listing (recursively by default)}
}
\description{
Get URLs of Files in Zipped ESRI-ascii-grid Format
}
