% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_monthly_data_for_berlin.R
\name{load_potential_evaporation_berlin_2}
\alias{load_potential_evaporation_berlin_2}
\title{New Version of Reading Potential Evaporation for Berlin}
\usage{
load_potential_evaporation_berlin_2(from, to)
}
\arguments{
\item{from}{optional. First month to be considered, as yyyymm string}

\item{to}{optional. Last month to be considered, as yyyymm string}
}
\value{
data frame with columns \code{file} (name of file downloaded from
DWD), \code{year} (year number as integer), \code{month number as integer},
\code{mean} (mean value), \code{sd} (standard deviation), \code{min}
(minimum value), \code{max} (maximum value), \code{n_values} (number of
considered values) of potential evaporation calculated for Berlin, Germany
}
\description{
This version uses publicly available shape files to determine the area that
is covered by Berlin. This function can easily be adopted to other cities!
}
