% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_precipitation_berlin.R
\name{load_precipitation_berlin}
\alias{load_precipitation_berlin}
\title{Load precipitation data from DWD server for Berlin area}
\usage{
load_precipitation_berlin()
}
\value{
data frame with columns \code{file} (name of file downloaded from
  DWD), \code{year} (year number as integer), \code{month number as integer},
  \code{mean} (mean value), \code{sd} (standard deviation), \code{min}
  (minimum value), \code{max} (maximum value) of potential evaporation
  calculated for Berlin, Germany
}
\description{
This function loads monthly precipitation values for Berlin, Germany,
from Deutscher Wetterdienst (DWD).
}
