% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_description.R
\name{open_description}
\alias{open_description}
\title{Open Description for DWD Record in Web Browser}
\usage{
open_description(url)
}
\arguments{
\item{url}{URL to DWD data file on ftp server}
}
\description{
This function looks for text or PDF files in the same folder as a DWD data
file, asks the user which file to open if there is more than one file and
tries to open the file in the web browser.
}
