% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{opendata_files}
\alias{opendata_files}
\title{Files Available on DWD's FTP Server}
\format{A data frame with 436687 rows and 2 variables:
\describe{
  \item{path}{URL of the path to the file on the FTP server}
  \item{file}{file name}
}}
\source{
\url{ftp://opendata.dwd.de/climate_environment/CDC/Announce_log_CDC_ftp.txt}
}
\usage{
opendata_files
}
\description{
URLs to files that are available for download at
\url{ftp://opendata.dwd.de/climate_environment/CDC/}. The file list was last
updated on 2019-06-01.
}
\keyword{datasets}
