% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radolan_raw_to_raster.R
\name{radolan_raw_to_raster}
\alias{radolan_raw_to_raster}
\title{Radolan raw to raster}
\usage{
radolan_raw_to_raster(rbi, version = 3)
}
\arguments{
\item{rbi}{rbi}

\item{version}{version (default: 3)}
}
\value{
???
}
\description{
Radolan raw to raster
}
\seealso{
\itemize{
\item \link{get_radolan_urls},
\item \link{extract_radolan_zip_files},
\item \link{download_radolan},
\item \link{list_extracted_radolan_files}.
}
}
