% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_asc_gz_file.R
\name{read_asc_file}
\alias{read_asc_file}
\title{Read Raster Data from .ASC File}
\usage{
read_asc_file(
  file,
  projection = readLines(default_projection_file()),
  dbg = TRUE
)
}
\arguments{
\item{file}{path to .asc file}

\item{projection}{projection string used in Radolan data}

\item{dbg}{logical indicating whether to show debug messages}
}
\description{
Read Raster Data from .ASC File
}
