% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_binary_radolan_file.R
\name{read_binary_radolan_file}
\alias{read_binary_radolan_file}
\title{Read Binary Radolan File}
\usage{
read_binary_radolan_file(path, set_projection_and_extent = TRUE)
}
\arguments{
\item{path}{path to binary Radolan file}

\item{set_projection_and_extent}{if \code{TRUE} (default), the projection
and extent of the raster object are set to the Radolan specific values,
returned by \code{kwb.dwd:::get_radolan_projection_string()} and
\code{get_radolan_extension()}, respectively.}
}
\description{
Read Binary Radolan File
}
