% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_daily_data_over_shape.R
\name{read_daily_data_over_shape}
\alias{read_daily_data_over_shape}
\title{Read daily data from DWD, mask region with given shape file}
\usage{
read_daily_data_over_shape(file, variable, from, to)
}
\arguments{
\item{file}{path to shape file .shp}

\item{variable}{currently, the following variables are supported: "evapo_p",
"evapo_r", "frost_depth", "soil_moist", "soil_temperature_5cm"}

\item{from}{first month as "yyyymm" string}

\item{to}{last month as "yyyymm" string}
}
\value{
data frame
}
\description{
Currently, only full months of data can be loaded, \code{from} and \code{to}
must be given
}
