% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_over_shape.R
\name{read_data_over_shape}
\alias{read_data_over_shape}
\title{Read data from DWD, mask region with given shape file or spatial object}
\usage{
read_data_over_shape(
  resolution,
  file,
  variable,
  from,
  to,
  quiet = FALSE,
  shape = NULL,
  use_sf = TRUE,
  ...
)
}
\arguments{
\item{resolution}{one of "daily", "monthly"}

\item{file}{path to shape file .shp}

\item{variable}{currently, the following variables are supported: "evapo_p",
"evapo_r", "frost_depth", "soil_moist", "soil_temperature_5cm" (daily),
"evapo_p" (monthly)}

\item{from}{first month as "yyyymm" string}

\item{to}{last month as "yyyymm" string}

\item{quiet}{passed to \link{download.file}}

\item{shape}{object of class sf or sf_layers or SpatialPolygonsDataFrame.
If \code{NULL} (the default), this object is obtained by calling
\code{kwb.dwd:::read_shape_file} on \code{file}.}

\item{use_sf}{passed to \code{kwb.dwd:::read_shape_file} if applicable}

\item{\dots}{further arguments passed to \code{kwb.dwd:::read_shape_file},
such as \code{drop_z = TRUE}, if applicable}
}
\value{
data frame
}
\description{
Read data from DWD, mask region with given shape file or spatial object
}
