% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_monthly_data_over_shape.R
\name{read_monthly_data_over_shape}
\alias{read_monthly_data_over_shape}
\title{Read monthly data from DWD, mask region with given shape file}
\usage{
read_monthly_data_over_shape(file, variable, from, to, ...)
}
\arguments{
\item{file}{path to shape file .shp}

\item{variable}{currently only "evapo_p" is supported}

\item{from}{first month as "yyyymm" string}

\item{to}{last month as "yyyymm" string}

\item{\dots}{further arguments passed to \code{kwb.dwd:::read_shape_file},
such as \code{drop_z = TRUE, use_sf = TRUE}}
}
\value{
data frame
}
\description{
Read monthly data from DWD, mask region with given shape file
}
