% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_relevant_years_radolan.R
\name{read_relevant_years_radolan}
\alias{read_relevant_years_radolan}
\title{Reads Relevant Radolan Data from rst Database}
\usage{
read_relevant_years_radolan(path, years)
}
\arguments{
\item{path}{path to rst database (folder)}

\item{years}{vector of year numbers for which to read Radolan data files}
}
\description{
This function reads in relevant radolan data from rst database based on used
input data
}
