% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zipped_esri_ascii_grid_from_url.R
\name{read_zipped_esri_ascii_grid_from_url}
\alias{read_zipped_esri_ascii_grid_from_url}
\title{Read Zipped ESRI-Ascii-Grid File from URL}
\usage{
read_zipped_esri_ascii_grid_from_url(url)
}
\arguments{
\item{url}{URL to zipped file in ESRI-ascii-grid format}
}
\value{
object of class "RasterLayer"
}
\description{
Read Zipped ESRI-Ascii-Grid File from URL
}
