% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_shapes.R
\name{select_shapes}
\alias{select_shapes}
\title{Interactively Configure Selection of SpatialPolygons}
\usage{
select_shapes(shapes)
}
\arguments{
\item{shapes}{list of SpatialPolygonsDataFrame objects, as e.g. returned by
\code{\link{get_shapes_of_germany}}}
}
\value{
list with elements \code{index} (index of SpatialPolygonsDataFrame in
  \code{shapes} list), \code{variable} (column in selected
  SpatialPolygonsDataFrame), \code{pattern} pattern (e.g. the name of a city)
  to be matched against the values in the selected column of the
  selected SpatialPolygonsDataFrame. This list simply describes how to select
  a SpatialPolygon from a list of SpatialPolygonsDataFrames
}
\description{
Interactively Configure Selection of SpatialPolygons
}
