% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_shapes.R
\name{select_variable}
\alias{select_variable}
\title{Let User Select a Column from a SpatialPolygonsDataFrame}
\usage{
select_variable(shape)
}
\arguments{
\item{shape}{object of class SpatialPolygonsDataFrame}
}
\value{
name of selected column
}
\description{
Let User Select a Column from a SpatialPolygonsDataFrame
}
