% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shapes_germany_0}
\alias{shapes_germany_0}
\title{Shapes of German Regions (0)}
\format{
Simple feature collection with 1 feature and 2 fields:
\itemize{
\item \code{COUNTRY},
\item \code{ID_0}.
}
}
\source{
\url{https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_DEU_shp.zip}
}
\usage{
shapes_germany_0
}
\description{
Object of class "sf" representing a "Simple feature collection".
}
\keyword{datasets}
