% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shapes_germany_1}
\alias{shapes_germany_1}
\title{Shapes of German Regions (1)}
\format{
Simple feature collection with 16 features and 11 fields:
\itemize{
\item \code{ID_0},
\item \code{COUNTRY},
\item \code{ID_1},
\item \code{NAME_1},
\item \code{VARNAME_1},
\item \code{NL_NAME_1},
\item \code{TYPE_1},
\item \code{ENGTYPE_1},
\item \code{CC_1},
\item \code{HASC_1},
\item \code{ISO_1}.
}
}
\source{
\url{https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_DEU_shp.zip}
}
\usage{
shapes_germany_1
}
\description{
Object of class "sf" representing a "Simple feature collection".
}
\keyword{datasets}
