% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_dir}
\alias{temp_dir}
\title{Path to Permanent Temporary Directory}
\usage{
temp_dir(..., template = NULL, create = TRUE, dbg = FALSE)
}
\arguments{
\item{\dots}{parts of the path after \verb{<TEMP_DIR>/R_kwb.dwd/}, passed to
\link{file.path}] where \verb{<TEMP_DIR>} is either the value of
environment variable \code{TEMP} (if set) or \code{TMP} (if set) or the
result of calling \link{tempdir}.}

\item{template}{optional. If given, it is assumed to be a path to a file. The
name of the file without file name extension is then used as folder name
below \verb{<TEMP_DIR>/R_kwb.dwd/}.}

\item{create}{logical indicating whether or not to create the folder if it
does not yet exist. Defaults to \code{TRUE}.}

\item{dbg}{logical indicating whether or not to print debug messages}
}
\value{
The function returns the path to the temporary folder specified.
}
\description{
Path to Permanent Temporary Directory
}
