% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_asc_gz_file.R
\name{unzip_asc_gz_file}
\alias{unzip_asc_gz_file}
\title{Download a .gz File and Unzip it}
\usage{
unzip_asc_gz_file(file, target_dir = tempdir())
}
\arguments{
\item{file}{path to asc.gz file to be unzipped or URL from which the file is
to be downloaded first}

\item{target_dir}{path to folder into which to (if applicable, download and)
unzip the file}
}
\value{
path to the unzipped file
}
\description{
Download a .gz File and Unzip it
}
