% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin_to_raster_file.R
\name{write_raster_to_file}
\alias{write_raster_to_file}
\title{Write raster to file}
\usage{
write_raster_to_file(rb, file, dbg = TRUE)
}
\arguments{
\item{rb}{raster data}

\item{file}{file path}

\item{dbg}{debug (TRUE/FALSE)}
}
\value{
write raster to file
}
\description{
Write raster to file
}
\seealso{
\itemize{
\item \link{convert_bin_to_raster_file}.
}
}
