% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_monthly_variable_for_region.R
\name{calculate_masked_grid_stats}
\alias{calculate_masked_grid_stats}
\title{Calculate Stats of Variable for Geographical Subset of a Grid}
\usage{
calculate_masked_grid_stats(matrices, geo_mask)
}
\arguments{
\item{matrices}{matrices}

\item{geo_mask}{"mask matrix" defining a geographical subset}
}
\value{
data frame with one row per matrix in \code{matrices} and columns
\itemize{
\item \code{file},
\item \code{year},
\item \code{month},
\item \code{mean},
\item \code{sd},
\item \code{min},
\item \code{max}
}
}
\description{
Calculate Stats of Variable for Geographical Subset of a Grid
}
