% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_or_download_shapes_germany.R
\name{check_or_download_shapes_germany}
\alias{check_or_download_shapes_germany}
\title{Check Local Availability or Download Shape Files}
\usage{
check_or_download_shapes_germany(quiet = FALSE)
}
\arguments{
\item{quiet}{passed if \code{TRUE} status messages are suppressed}
}
\value{
path to folder containing shape files
}
\description{
This function checks if shape files for Germany are available in the folder
`\%TEMP\%/R_kwb.dwd/gadm40_DEU_shp`. If not, the shape files are downloaded
from \url{https://gadm.org} (URL to zip-file:
\url{https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_DEU_shp.zip})
and unpacked into that folder.
}
