% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_or_download_shapes_germany.R
\name{check_or_download_shapes_germany}
\alias{check_or_download_shapes_germany}
\title{Check Local Availability or Download Shape Files}
\usage{
check_or_download_shapes_germany(
  url = "https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_DEU_shp.zip",
  quiet = FALSE,
  timeout = 60
)
}
\arguments{
\item{url}{URL to publicly available zip file containing shape files for
Germany. Default:
\url{https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_DEU_shp.zip}}

\item{quiet}{passed if \code{TRUE} status messages are suppressed}

\item{timeout}{timeout in seconds}
}
\value{
path to folder containing shape files
}
\description{
This function checks if shape files for Germany are available in the folder
\verb{\\\%TEMP\\\%/R_kwb.dwd/shapes_germany}. If not, the shape files are
downloaded from the URL given in \code{url} and unpacked into that folder.
}
\seealso{
\itemize{
\item \link{get_shapes_of_germany},
\item \link{get_example_grid_germany},
\item \link{download_radolan}.
}
}
