% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_radolan_radial.R
\name{crop_radolan_radial}
\alias{crop_radolan_radial}
\title{Crop Radolan Data}
\usage{
crop_radolan_radial(
  radolan,
  longitude,
  latitude,
  radius = 10,
  as_data_frame = TRUE
)
}
\arguments{
\item{radolan}{radolan data as returned by \link{read_relevant_years_radolan}}

\item{longitude}{longitude}

\item{latitude}{latitude}

\item{radius}{distance from point (latitude/longitude) in km}

\item{as_data_frame}{if \code{TRUE}, a data frame with columns
\code{SAMPLE_DATE} and \code{rain_mean} is returned, otherwise the raster
object returned by \link[raster:crop]{raster::crop}}
}
\description{
Crop Radolan Data
}
\seealso{
\itemize{
\item \link{get_radolan_urls},
\item \link{extract_radolan_zip_files},
\item \link{radolan_raw_to_raster},
\item \link{download_radolan}.
}
}
