% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_radolan_zip_files.R
\name{extract_radolan_zip_files}
\alias{extract_radolan_zip_files}
\title{Extract radolan zip files}
\usage{
extract_radolan_zip_files(radolan_dir)
}
\arguments{
\item{radolan_dir}{path to directory containing zipped Radolan files (ending
in .tar.gz)}
}
\value{
Nothing. Side effect: Unzipped files in subfolder "bin" below
\code{radolan_dir}.
}
\description{
Extract radolan zip files
}
\seealso{
\itemize{
\item \link{list_extracted_radolan_files}.
}
}
