% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_binary_radolan_file.R
\name{get_radolan_extension}
\alias{get_radolan_extension}
\title{Get Radolan Extension}
\usage{
get_radolan_extension()
}
\value{
default radolan extension
}
\description{
Get Radolan Extension
}
\examples{
get_radolan_extension()

}
\seealso{
\itemize{
\item \link{get_radolan_urls},
\item \link{get_radolan_projection_string}.
}
}
