% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radolan_urls.R
\name{get_radolan_urls}
\alias{get_radolan_urls}
\title{Get URLs to Available Radolan Files}
\usage{
get_radolan_urls(
  start_daily = "2006-10",
  start_hourly = "2005-06",
  end_daily = format(format = "\%Y-\%m", lubridate::rollback(Sys.Date(), roll_to_first =
    TRUE)),
  end_hourly = end_daily,
  ...
)
}
\arguments{
\item{start_daily}{month string (yyyy-mm) of first daily record.
Default: "2006-10"}

\item{start_hourly}{month string (yyyy-mm) of first hourly record.
Default: "2005-06"}

\item{end_daily}{month string (yyyy-mm) of last daily record. Defaults to the
current month.}

\item{end_hourly}{month string (yyyy-mm) of last hourly records. Defaults to
\code{end_daily}.}

\item{\dots}{further arguments passed to \code{kwb.dwd:::get_radolan_url},
such as \code{ftp_root}}
}
\value{
list with "daily_historical_urls" and "hourly_historical_urls"
}
\description{
Get URLs to available radolan files on the DWD FTP server below this base
address: \url{ftp://opendata.dwd.de/climate_environment/CDC/grids_germany}.
The user can choose between daily records and hourly records that are located
at different paths on the server. The paths to the files are not read from
the FTP server but generated, given the knowledge of where the files should
reside and how they are expected to be named.
}
\examples{
# Get all expected URLs
urls <- kwb.dwd::get_radolan_urls()

# Show the first URLs of each category (daily, hourly)
head(urls$daily_historical_urls)
head(urls$hourly_historical_urls)

# Set the start and end months
urls <- kwb.dwd::get_radolan_urls(
  start_daily = "2017-01", end_daily = "2017-03"
)

urls$daily_historical_urls

}
\seealso{
\itemize{
\item \link{get_dwd_urls_metadata},
\item \link{get_radolan_extension},
\item \link{get_radolan_projection_string}.
}
}
