% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapes_of_germany.R
\name{get_shapes_of_germany}
\alias{get_shapes_of_germany}
\title{Provide List of SpatialPolygonsDataFrame for Germany}
\usage{
get_shapes_of_germany(recreate = FALSE)
}
\arguments{
\item{recreate}{logical. If \code{TRUE} the required shape files are freshly
downloaded from \url{https://gadm.org/}, transformed to the projection used
in DWD data files and stored locally in an RData file. The default is
\code{FALSE}, i.e. the required data are read form the locally stored RData
file, provided that the file exists.}
}
\description{
Provide List of SpatialPolygonsDataFrame for Germany
}
