% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_url.R
\name{list_url}
\alias{list_url}
\title{List Files or Folders in a FTP Directory}
\usage{
list_url(url, recursive = FALSE, max_depth = NA, ..., depth = 0,
  full_info = FALSE)
}
\arguments{
\item{url}{URL to FTP server, including "ftp://"}

\item{recursive}{logical indicating whether to list files in all
subdirectories (default: \code{FALSE})}

\item{max_depth}{maximum folder depth to consider when
\code{recursive = TRUE}}

\item{\dots}{arguments passed to \code{kwb.dwd:::try_to_get_url}, such as
\code{n_trials}, \code{timeout}, or \code{sleep_time}}

\item{depth}{for start depth when \code{recursive = TRUE}}

\item{full_info}{if \code{TRUE}, not only the path and filename are returned
but also the file properties. The default is \code{FALSE}.}
}
\description{
List Files or Folders in a FTP Directory
}
