% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_monthly_data_for_berlin.R
\name{load_potential_evaporation_berlin}
\alias{load_potential_evaporation_berlin}
\title{Load monthly potential evaporation for Berlin from DWD}
\usage{
load_potential_evaporation_berlin(from = NULL, to = NULL)
}
\arguments{
\item{from}{optional. First month to be considered, as yyyymm string}

\item{to}{optional. Last month to be considered, as yyyymm string}
}
\value{
data frame with columns \code{file} (name of file downloaded from
  DWD), \code{year} (year number as integer), \code{month number as integer},
  \code{mean} (mean value), \code{sd} (standard deviation), \code{min}
  (minimum value), \code{max} (maximum value), \code{n_values} (number of
  considered values) of potential evaporation calculated for Berlin, Germany
}
\description{
This function loads monthly potential evaporation for Berlin, Germany,
from Deutscher Wetterdienst (DWD).
}
