% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radolan_raw_to_raster.R
\name{radolan_raw_to_raster}
\alias{radolan_raw_to_raster}
\title{Radolan raw to raster}
\usage{
radolan_raw_to_raster(rbi, version = 3)
}
\arguments{
\item{rbi}{rbi}

\item{version}{version (default: 3)}
}
\value{
???
}
\description{
Radolan raw to raster
}
