% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_asc_gz_file.R
\name{read_asc_gz_file}
\alias{read_asc_gz_file}
\title{Read Zipped ESRI-Ascii-Grid File (from URL)}
\usage{
read_asc_gz_file(file, url = NULL)
}
\arguments{
\item{file}{path to zipped file in ESRI-ascii-grid format (.asc.gz)}

\item{url}{optional. URL to zipped file in ESRI-ascii-grid format (.asc.gz)}
}
\value{
object of class "RasterLayer"
}
\description{
Read Zipped ESRI-Ascii-Grid File (from URL)
}
\seealso{
\itemize{
\item \link{read_binary_radolan_file},
\item \link{unzip_asc_gz_file}.
}
}
