% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_binary_radolan_file.R
\name{read_binary_radolan_file}
\alias{read_binary_radolan_file}
\title{Read Binary Radolan File}
\usage{
read_binary_radolan_file(
  path,
  set_projection_and_extent = TRUE,
  consider_flags = FALSE
)
}
\arguments{
\item{path}{path to binary Radolan file}

\item{set_projection_and_extent}{if \code{TRUE} (default), the projection
and extent of the raster object are set to the Radolan specific values,
returned by \code{kwb.dwd::get_radolan_projection_string} and
\code{kwb.dwd::get_radolan_extension}, respectively.}

\item{consider_flags}{logical. Should the flags be considered? If
\code{TRUE}, values where the "invalid" or "clutter" flag is set are set to
\code{NA} and values where the "negative" flag is set are negated. The
default is \code{TRUE} (for compatibility reasons) but this should change
in future!}
}
\description{
Read Binary Radolan File
}
\seealso{
\itemize{
\item \link{read_asc_gz_file},
\item \link{read_relevant_years_radolan}.
}
}
