% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shapes_germany_1}
\alias{shapes_germany_1}
\title{Shapes of German Regions (1)}
\format{
Simple feature collection with 16 features and 11 fields:
\code{ID_0}, \code{COUNTRY}, \code{ID_1}, \code{NAME_1}, \code{VARNAME_1},
\code{NL_NAME_1}, \code{TYPE_1}, \code{ENGTYPE_1}, \code{CC_1},
\code{HASC_1}, \code{ISO_1}
}
\source{
\url{https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_DEU_shp.zip}
}
\usage{
shapes_germany_1
}
\description{
Object of class "sf" representing a "Simple feature collection".
}
\keyword{datasets}
