% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shapes_germany_2}
\alias{shapes_germany_2}
\title{Shapes of German Regions (2)}
\format{
Simple feature collection with 403 features and 12 fields:
\itemize{
\item \code{ID_0},
\item \code{COUNTRY},
\item \code{NAME_1},
\item \code{NL_NAME_1},
\item \code{ID_2},
\item \code{NAME_2},
\item \code{VARNAME_2},
\item \code{NL_NAME_2},
\item \code{TYPE_2},
\item \code{ENGTYPE_2},
\item \code{CC_2},
\item \code{HASC_2}.
}
}
\source{
\url{https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_DEU_shp.zip}
}
\usage{
shapes_germany_2
}
\description{
Object of class "sf" representing a "Simple feature collection".
}
\keyword{datasets}
