% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.R
\name{unzip_asc_gz_file}
\alias{unzip_asc_gz_file}
\title{Download a .gz File and Unzip it}
\usage{
unzip_asc_gz_file(file, target_dir = tempdir())
}
\arguments{
\item{file}{path to asc.gz file to be unzipped or URL from which the file is
to be downloaded first}

\item{target_dir}{path to folder into which to (if applicable, download and)
unzip the file}
}
\value{
path to the unzipped file
}
\description{
Download a .gz File and Unzip it
}
\seealso{
\itemize{
\item \link{read_binary_radolan_file},
\item \link{convert_bin_to_raster_file},
\item \link{read_asc_gz_file},
\item \link{write_raster_to_file}.
}
}
